pyproject.toml
[tool.poetry]
name = "voyage"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
jupyterlab = "^4.0.9"
voyageai = "^0.1.3"
pandas = {extras = ["output-formatting", "plot"], version = "^2.1.4"}
pyarrow = "^14.0.1"
python-decouple = "^3.8"
python-dotenv = "^1.0.0"
polars = "^0.20.2"
openai = "^1.6.1"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-cov = "^4.1.0"
syrupy = "^4.6.0"

[tool.pytest.ini_options]
addopts = "--strict-markers"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
]

[tool.coverage.run]
source = ["voyage"]
omit = ["*/__init__.py"]

[tool.coverage.report]
exclude_lines = ["raise NotImplementedError"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
